LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-anchor-positioning-001-expected.txt


FAIL CSS Content Visibility: auto prevents existing anchors to unskipped content
from becoming skipped assert_equals: #positioned should be anchored because #anchor is on-screen. expected 209 but got 8
